Change quantity fields from int to decimal for accuracy
Updated quantity-related variables and parsing logic from int to decimal in chalan and print modules to support fractional quantities. Also adjusted table column widths in Update_Designation.aspx for improved layout.

+++ Bill_Software/corporate/business/app/add_chalan.aspx.cs
@@ -206,8 +206,8 @@ namespace Bill_Software.corporate.business.app
                             //}
 
                             string Qty = ((TextBox)gd_Quotation.Rows[i].FindControl("Qty")).Text?.Trim() ?? "";
-                            int quantity = 0;
-                            if (!int.TryParse(Qty, out quantity))
+                            decimal quantity = 0;
+                            if (!decimal.TryParse(Qty, out quantity))
                             {
                                 // optionally: show error next to row or skip
                                 quantity = 0;
