+++ Bill_Software/corporate/business/print/NewChhalan.aspx.cs
@@ -22,7 +22,7 @@ namespace Bill_Software.corporate.business.print
         DataTable dtRep = new DataTable();
         DataTable dtChadd = new DataTable();
         StringBuilder strProduct = new StringBuilder();
-        public int TQ = 0;
+        public decimal TQ = 0;
 
         CultureInfo cultureInfo = Thread.CurrentThread.CurrentCulture;
         protected void Page_Load(object sender, EventArgs e)
@@ -187,7 +187,7 @@ namespace Bill_Software.corporate.business.print
                     string Product_id = dtChPro.Rows[i]["Product_id"].ToString();
                     string HSN = dtChPro.Rows[i]["Product_code"].ToString();
                     string Product_name = dtChPro.Rows[i]["Product_name"].ToString();
-                    int Quantity = Convert.ToInt32(dtChPro.Rows[i]["Quantity"]);
+                    decimal Quantity = Convert.ToDecimal(dtChPro.Rows[i]["Quantity"]);
                     string Unit = dtChPro.Rows[i]["Unit"].ToString();
                     string Specification = dtChPro.Rows[i]["specification"].ToString();
                     string ItemNo = dtChPro.Rows[i]["ItemNo"].ToString();
